requirements.txt
beautifulsoup4==4.12.3
pandas
numpy
python-dotenv==1.0.1
python_dateutil==2.8.2
selenium==4.24.0
webdriver_manager==4.0.1
flask 
